feat: improve ui in home page and code editor page
