+++ lib/features/profile/presentation/widgets/leetcode_profile.dart
@@ -149,7 +149,7 @@ class LeetcodeProfile extends HookWidget {
           .map(
             (e) => Padding(
               padding: const EdgeInsets.symmetric(
-                horizontal: 16.0,
+                horizontal: 8.0,
                 vertical: 8.0,
               ),
               child: e,
